populate the content in case the exec in p4.Run() fails
The error returned in the current code is unhelpful in some cases e.g.
when the password has expired. Currently the error returned by p4.Run()
is just "exit status 1" with empty content. This commit updates the
p4.Run() function to also populate the results returned to the caller
as the details about the p4 ticket being expired is printed by the p4
binary in stdout.